Module1 Part completed
Could you change the filename for me? Also, I completed all of the first module of the project. I need to test the rest of the instructions and operands. Overall, it works, and I wanted also to see how we want to handle specific errors such as an invalid jump(out of bounds error), division by 0, invalid zero jump(Out of bounds error) should be handled. Like should the whole program halt because of them or should they continue the program?